# expression: implement vectorized evaluation for `builtinMinuteSig`