<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
This PR implements vectorized `builtinMinuteSig`.
Issue: #12103

### What is changed and how it works?
```
$ go test ./expression/ -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args "builtinMinuteSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-4           1000000000               0.0133 ns/op          0 B/op             0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinMinuteSig-VecBuiltinFunc-4              178981              6528 ns/op          0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinMinuteSig-NonVecBuiltinFunc-4            44202             27054 ns/op          0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.516s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test