@rogaps 2019-10-23
Great. Let me finish the rests.